requirements.txt
streamlit>=1.28.0
pypdf>=3.16.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.4
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.1.0
transformers>=4.36.0
